[SG-42] move & rename: post 하위 패키지로 이동, 파일 및 함수명 변경
